requirements.txt
aiogram
yt-dlp
python-dotenv
aiohttp
aiofiles
